+++ R/generics.R
@@ -66,15 +66,26 @@ as_raw <- function(x, ...) {
   UseMethod("as_raw")
 }
 
-#' @title Number of Dimensions
+#' @title Number of Axes
 #' @description
-#' Get the number of dimensions of a tensor.
+#' Get the number of axes of a tensor.
 #' @param x (any)\cr
 #'   Object.
 #' @return (integer)
 #' @export
-ndims <- function(x) {
+naxes <- function(x) {
   length(shape(x))
+}
+
+#' @title Axes
+#' @description
+#' Get the axes of a tensor, i.e. `seq_len(naxes(x))`.
+#' @param x (any)\cr
+#'   Object.
+#' @return (`integer()`)
+#' @export
+axes <- function(x) {
+  seq_len(naxes(x))
 }
 
 #' @title Number of Elements